+++ Arbiter.App/ViewModels/Client/ClientManagerViewModel.cs
@@ -1,6 +1,9 @@
-﻿using System.Collections.ObjectModel;
+﻿using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.Linq;
+using System.Threading;
 using Arbiter.Net;
+using Avalonia.Threading;
 
 namespace Arbiter.App.ViewModels.Client;
 
@@ -8,6 +11,9 @@ public class ClientManagerViewModel : ViewModelBase
 {
     private readonly ProxyServer _proxyServer;
 
+    private readonly Lock _clientLock = new();
+    private readonly Dictionary<int, ClientViewModel> _clientMap = [];
+    
     public ObservableCollection<ClientViewModel> Clients { get; } = [];
 
     public ClientManagerViewModel(ProxyServer proxyServer)
@@ -15,6 +21,7 @@ public class ClientManagerViewModel : ViewModelBase
         _proxyServer = proxyServer;
 
         _proxyServer.ClientAuthenticated += OnClientAuthenticated;
+        _proxyServer.ClientLoggedIn += OnClientLoggedIn;
         _proxyServer.ClientRedirected += OnClientRedirected;
         _proxyServer.ClientDisconnected += OnClientDisconnected;
     }
@@ -24,8 +31,25 @@ public class ClientManagerViewModel : ViewModelBase
         var connection = e.Connection;
         var client = new ClientViewModel(connection) { Id = connection.Id, Name = connection.Name ?? string.Empty };
 
+        // Start listening for packets and updating state
         client.Subscribe();
-        Clients.Add(client);
+
+        using var _ = _clientLock.EnterScope();
+        _clientMap[connection.Id] = client;
+    }
+
+    private void OnClientLoggedIn(object? sender, ProxyConnectionEventArgs e)
+    {
+        using var _ = _clientLock.EnterScope();
+        var client = _clientMap.GetValueOrDefault(e.Connection.Id);
+
+        if (client is null)
+        {
+            return;
+        }
+
+        // We are fully logged in and can display this client in the list
+        Dispatcher.UIThread.Invoke(() => Clients.Add(client));
     }
 
     private void OnClientRedirected(object? sender, ProxyConnectionRedirectEventArgs e)
@@ -38,9 +62,14 @@ public class ClientManagerViewModel : ViewModelBase
         var client = Clients.FirstOrDefault(c => c.Id == e.Connection.Id);
         client?.Unsubscribe();
 
-        if (client is not null)
+        if (client is null)
         {
-            Clients.Remove(client);
+            return;
         }
+        
+        Dispatcher.UIThread.Invoke(() => Clients.Remove(client));
+        
+        using var _ = _clientLock.EnterScope();
+        _clientMap.Remove(client.Id);
     }
 }